+++ src/appliance-onboarding-script/appliance_setup/pkgs/_appliance_setup.py
@@ -4,7 +4,7 @@ import logging
 from ._exceptions import AzCommandError, InvalidOperation, ProgramExit, ArmFeatureNotRegistered, \
     ClusterExtensionCreationFailed, ArmProviderNotRegistered
 from ._az_cli import az_cli
-from ._checks_and_validations import _wait_until_appliance_is_in_running_state
+from ._azure_resource_validations import _wait_until_appliance_is_in_running_state
 from ._utils import TempChangeDir, confirm_prompt, decode_base64, delete_empty_sub_dicts, delete_unassigned_fields, get_value_using_path_in_dict, set_value_using_path_in_dict, create_dir_if_doesnot_exist, get_vm_snapshot_name
 from . import _templates as templates
 from shutil import copy
@@ -277,10 +277,11 @@ class ApplianceSetup(object):
 
         res = None
         if op == 'create':
-
-            _wait_until_appliance_is_in_running_state(config, "Create K8s Extension")
-
+            
             logging.info('Creating VMware extension...')
+
+            _wait_until_appliance_is_in_running_state(config)
+
             res, err = az_cli('k8s-extension', 'create', '--debug',
                 '-n', name,
                 '-g', rg,
