+++ src/appliance-onboarding-script/appliance_setup/pkgs/_arcvmware_resources.py
@@ -5,7 +5,7 @@ import re
 from pkgs._utils import safe_quote_string
 from ._az_cli import az_cli
 from ._exceptions import AzCommandError
-from ._checks_and_validations import _wait_until_appliance_is_in_running_state
+from ._azure_resource_validations import _wait_until_appliance_is_in_running_state
 
 class ArcVMwareResources(object):
 
@@ -29,9 +29,10 @@ class ArcVMwareResources(object):
     def _create_cl(self, appliance_id, extension_id) -> str:
         config = self._config
 
-        _wait_until_appliance_is_in_running_state(config, "Create Custom Location")
-
         logging.info('Creating Custom Location...')
+
+        _wait_until_appliance_is_in_running_state(config)
+
         location = config['location']
         rg = config['resourceGroup']
         name = config['customLocationAzureName']
@@ -64,9 +65,10 @@ class ArcVMwareResources(object):
     def _connect_vcenter(self, custom_location_id: str):
         config = self._config
 
-        _wait_until_appliance_is_in_running_state(config, "Connect vCenter")
-
         logging.info('Connecting vCenter...')
+
+        _wait_until_appliance_is_in_running_state(config)
+        
         location = config['location']
         rg = config['resourceGroup']
         name = config['nameForVCenterInAzure']
